--- a/src/torematrix/ui/components/element_list/models/__init__.py
+++ b/src/torematrix/ui/components/element_list/models/__init__.py
@@ -1,0 +1,10 @@
+"""
+Element List Models
+
+Data models for hierarchical element representation.
+"""
+
+from .tree_model import ElementTreeModel
+from .tree_node import TreeNode
+
+__all__ = ['ElementTreeModel', 'TreeNode']