… AGENT_1_INLINE.md
… AGENT_2_INLINE.md
… AGENT_3_INLINE.md
… AGENT_4_INLINE.md
… INLINE_COORDINATION.md
… ISSUE_23_AGENT_PROMPTS.md
… src/torematrix/ui/components/element_list/interactions/selection.py
… src/torematrix/ui/components/element_list/models/__pycache__/tree_node.cpython-312.pyc
… src/torematrix/ui/components/element_list/models/tree_model.py
… src/torematrix/ui/viewer/tools/history.py
… src/torematrix/ui/viewer/tools/hit_testing.py
… src/torematrix/ui/viewer/tools/manager.py
… src/torematrix/ui/viewer/tools/persistence.py
… src/torematrix/ui/viewer/tools/snapping.py
… tests/unit/ui/viewer/tools/test_optimization.py